Move 3.2 upgrade tasks into separate file
