+++ lib/alchemy/upgrader/three_point_two.rb
@@ -1,32 +1,4 @@
-require 'thor'
-
-class Alchemy::Upgrader::ThreePointTwoTask < Thor
-  include Thor::Actions
-
-  no_tasks do
-    def patch_acts_as_taggable_on_migrations
-      sentinel = /def self.up/
-
-      aato_file = Dir.glob('db/migrate/*_acts_as_taggable_on_migration.*.rb').first
-      if aato_file
-        inject_into_file aato_file,
-          "\n    # inserted by Alchemy CMS upgrader\n    return if table_exists?(:tags)\n",
-          { after: sentinel, verbose: true }
-      end
-
-      aato_file = Dir.glob('db/migrate/*_add_missing_unique_indices.*.rb').first
-      if aato_file
-        inject_into_file aato_file,
-          "\n    # inserted by Alchemy CMS upgrader\n    return if index_exists?(:tags, :name)\n",
-          { after: sentinel, verbose: true }
-      end
-    end
-
-    def inject_seeder
-      append_file "./db/seeds.rb", "Alchemy::Seeder.seed!\n"
-    end
-  end
-end
+require_relative 'tasks/three_point_two_task'
 
 module Alchemy
   module Upgrader::ThreePointTwo
@@ -35,14 +7,14 @@ module Alchemy
     def upgrade_acts_as_taggable_on_migrations
       desc 'Installs acts_as_taggable_on migrations.'
       if !`bundle exec rake railties:install:migrations FROM=acts_as_taggable_on_engine`.empty?
-        Alchemy::Upgrader::ThreePointTwoTask.new.patch_acts_as_taggable_on_migrations
+        Alchemy::Upgrader::Tasks::ThreePointTwoTask.new.patch_acts_as_taggable_on_migrations
       end
       `bundle exec rake db:migrate`
     end
 
     def inject_seeder
       desc 'Add Alchemy seeder to `db/seeds.rb` file.'
-      Alchemy::Upgrader::ThreePointTwoTask.new.inject_seeder
+      Alchemy::Upgrader::Tasks::ThreePointTwoTask.new.inject_seeder
     end
 
     def alchemy_3_2_todos
